Updated search method.
Before, was using the find command in the music library directory, as well as
the current directory. If you were in a particularly high directory, then when
searching in the current directory, it would take ages.

This was fixed to only search the top level of the current directory, i.e. a
maxdepth of 1.